+++ internal/site/src/components/charts/container-chart.tsx
@@ -1,13 +1,13 @@
-import { Area, AreaChart, CartesianGrid, YAxis } from "recharts"
-import { type ChartConfig, ChartContainer, ChartTooltip, ChartTooltipContent, xAxis } from "@/components/ui/chart"
-import { memo, useMemo } from "react"
-import { cn, formatShortDate, chartMargin, toFixedFloat, formatBytes, decimalString } from "@/lib/utils"
 // import Spinner from '../spinner'
 import { useStore } from "@nanostores/react"
+import { memo, useMemo } from "react"
+import { Area, AreaChart, CartesianGrid, YAxis } from "recharts"
+import { type ChartConfig, ChartContainer, ChartTooltip, ChartTooltipContent, xAxis } from "@/components/ui/chart"
+import { ChartType, Unit } from "@/lib/enums"
 import { $containerFilter, $userSettings } from "@/lib/stores"
+import { chartMargin, cn, decimalString, formatBytes, formatShortDate, toFixedFloat } from "@/lib/utils"
 import type { ChartData } from "@/types"
 import { Separator } from "../ui/separator"
-import { ChartType, Unit } from "@/lib/enums"
 import { useYAxisWidth } from "./hooks"
 
 export default memo(function ContainerChart({